Redesign lookup infrastructure
Originally, lookups were intended to be tables, so there was a need to attach
an instance of the table to a lookup, which is not possible in pure bytecode.
The solution was to inject the instances after the generated code was
instantiated. This is sub-optimal for two reasons: the underlying object might
change its behaviour if it reloads its configuration and simple functions still
required instance injection.

This change makes lookups behave more like actions and constants: pure byte
code generation. Functions that do not require instances can simply generate
the apprpropriate method calls. In the case where an instance needs to be
tracked, INVOKE DYNAMIC is used. `LookupForInstance` holds a cache of all
“tables” or other instances and gives each on a unqiue key. It then generate
bytecode that calls INVOKE DYNAMIC on this unique key, allowing the instances
to connected at runtime in a less cumbersome way.