Change bases mask parser/packer
When modifying the bases mask objects, the input and output formats were the
same so it was easy to repack the same information. However, for new mask
objects, the `ordinal` had to be assigned even though it was discarded. This
was confusing in the olive, so the goal here is to make the input and output
formats logical, though different.