Attempt to perturb SFTP bug out of existence
There is a bug where the SFTP connection will get itself in a bad state and
throw errors when doing things. This is might due to some kind of
multi-threaded inteference, so ensure everything that access the connection is
synchronised. Upgrade the library in case it's a bug that's been fixed.
Finally, force the connection to be reset if it throws an exception, though
exceptions are thrown for "normal" errors (_e.g._ permissions problems).