Allow empty lists
Previously, empty lists weren't allowed because the type couldn't be
determined. This adds a limited type inference to allow an empty list type that
can be merged with a non-empty list. Empty lists cannot be directly iterated.

External entities can request empty lists (though this is silly to insist) and
empty lists can be used in places where a list is requested.