Disallow empty list as a parameter type for soundness
This prevents the empty list and empty optional type from being specified as a
parameter since it creates a hole in the type system:

```
Function destroy_everything([]x) x;

...
 For z In (If True Then destory_everything(["a"]) Else [1]):
  Reduce (a = 0) a + x
...
```

This will type check but explode at runtime. This is because the empty list
will unify with different inner types. The easiest solution is to prevent the
empty list as a parameter. If the empty list is returned, then it can be
confidently unified with any other type, even if it is the return of a
user-defined function.

All this applies to optional, but there is no way to write the empty optional
in a user-defined type.