Use IUSAttribute to store and compare signatures
JIRA: GP-1872

Attempt to store Shesmu signatures in Niassa as IUS annotations. This change
introduces copies of the CellRanger and Files LIMS key collections that provide
signatures. The other formats work as they did originally.

When matching against workflows is done, records which would be marked stale
but have identical signatures are considered fixable and LIMS key can be
updated. If no signatures are present and the record is fresh, signatures will
be automatically added.