Make simulation dashboard importer pluggable
Currently, the simulation dashboard can import actions in the format of
`/actions` and the `niassawf` format. Additional changes were necessary for new
Niassa features. This moves the import logic into the plugins and then supports
the new Niassa features:

- support `userAnnotations`
- support new LIMS key types and keep them automatically synchronised

This also changes the export format for actions into the `/actions` format.