Allow using signatures in LeftJoin
Signatures were only allowed on the main flow of data in an olive, but since
the data in a `LeftJoin` is unmanipulated, the signatures should be valid on
that data. This came about when trying to do a self-left join in `cerberus_fp`
which needed to combine signatures from both the outer an inner parts. This
means that all `LeftJoin` clauses must be proceeded by a reshaping clause that
drops or there will be a collision; in practice, this is already true for all
olives in use.