Combine JSON packing and unpacking
This decided which type to use before type variable resolution was applied. This mean the following:

```
TypeAlias foo json;

x = 3;

...
  x As json == x As foo
```

would produce a type error because `x As json` would produce the type `json`,
since it thinks is packing, and `x As foo` would produce the type `json?`,
since it thinks its unpacking.