Make JIRA suck marginally less
If you try to create a ticket in JIRA and aren't logged in, it will let you
fill out the form and then complain that reporter is missing. This forces all
calls to JIRA to go through a login redirect, which will be a no-op if all
ready logged in. Thanks, I hate it.