Allow exporting and sharing Define olives
This allows exporting `Define` olives such that they might be consumed by other
script. Doing this requires a very messy understanding of two data formats
being the same, so it generates 3 outputs for each olive defined:

- a single function to actually call the olive on a stream
- a function for each variable in the output stream to access the data since
  the caller will not have direct access to the methods that generate the
  output
- a function for every variable in the input data format to decide if it should
  be part of a signature or not

Typically, Shesmu has considered the stream to be in one of two states: pure
(_i.e._, the input data format) or transformed (_e.g._, the output of
grouping). This change introduces a new distinction: almost pure. An almost
pure stream is one that resembles the original data format but may have
additional fields added by `Flatten` or `Require`. This means that for the
purposes of generating signatures, the downstream operations should consider it
pure, but the actual objects output are _not_ the same as the input format.