Include INI parameters in Niassa action equality
There has been a lot of confusion about deduplication of actions after changing
INI parameters in the olive. At the time, this was an intentional design choice
to cope with the way that matching workflow runs against actions worked. Since
that matching algorithm has changed and there are more tools for interrogating
the output of matches, having the INI parameters act the way they do in Shesmu
is confusing. There was also a performance concern, but workflow run matching
is now sufficiently performant that if an olive suddenly produces another
complete set of actions, it won't cause load issues.