Create namespaces
This create namespaces to separate the names of actions, constants, functions,
and signatures. Creates a separate namespace for different plugin types. Some
related plugins are grouped into the same namespace. I've also reserved several
root namespaces for now and future use. In particular, the built-in
functionality is in `std` and scripts can export into `olive`.