Fix bad bytecode using signatures after `Flatten`
The `Flatten` (and `Require`) operation works by creating a proxy class with
two fields: the original value and the unrolled value and creating delegate
methods to the original value. Signatures do not exist as real methods on the
input data, so an incorrect delegate was created. This expands each signature
to a real value at the time of proxy construction and stores it as a field in
the proxy.