Change signature handling in `Define` olives
The changes signatures in `Define` olives to be passed as a single
`SignatureAccessor` object rather than individually and to copy this into
lambdas rather than storing this information in a field. The reason for this is
as a prerequisite to make `Export Define` possible. If a `Define` olive is
exported, so two different scripts might call it simultaneously, so it cannot
store state in a shared place. Further more, since signatures can be redefined
by plugins, putting the signatures in the method will make incompatible method
signatures if new signatures are found.