Replace dropdown initial with a predicate
There was causing a problem on the _Actions_ page where the initial was an
array and since `==` compares arrays for reference equality, the initial value
was never being selected. This changes it to use a predicate instead of
reference equality and makes sure to only update the target model once even if
the predicate returns true multiple times.