Fix bug where hidden histograms aren't rendered
Drawing to a canvas that is not currently displayed does not render anything to
the canvas. This lead to the situation where trigging a data refresh while on
another tab would result in blank histograms. This adds a `reveal` function
that gets called when a tab is selected and triggers a redraw of this histogram
at that time.