Create a Check pragma
This is a new pragma that allows stopping a script from running based on the
input data. It performs an aggregation of data and then evaluates an expression
with that aggregation. If the aggregation fails (_e.g._, Univalued), then it
does not run the test but stops execution.

The implementation is a bit messy and splits the core bytecode state handling
logic `Renderer` into subclasses depending on where the data stream is stored
(has none, is in argument, or in a local variable).