Create intersection join operations
This creates two new join operations that join on set intersections rather than
individual keys. The goal of this is to be able to match records again
hierarchically. The idea would be to do something like:

    Group By workflow_run_accession Into output_files = List accession
    LeftIntersectionJoin output_files To cerberus_fp(input_files)
      ...

This would allow finding files that use some of the output of another workflow
allowing building one layer of relationships between workflows.