Streamline UI internals
Currently, the UI has a bunch of things that get tracked out-of-band: the
reveal functions needed to trigger canvas drawing and the custom find handlers.
This changes the UI so both of those things are handled inline, simplifying
use. It also redesigns the drop down menus to use the same popup handlers as
everything else rather than CSS hover stuff.