Allow ignoring unused variables in simulation
Requiring all variables be used is cumbersome when developing; this change
allows variables to be unused during development. It also changes the
simulation UI to remember whether to wait for fresh data or use cached data and
presents all of these options in a more consistent way. It also drops the
`/checkhtml` endpoint, which was a sadder version of the simulator that is no
longer used.