Fix table menu filter lifetime
The table menu used on the _Olives_ page has incorrectly scoped state. The
filter box is recycled between dropdowns of the menu, but the effects of the
filter are not, causing it to look like it should be filtered even though it
isn't. This adjust the scope of that information to disappear when the dropdown
is closed.