Improve type safety of algebraic types comparisons
When doing a comparison (`==`, `!=`) of an algebraic type, `FOO == BAR` should
be an invalid comparison, but because they are both ADT, they are considered
mergable (`Imyhat.isSame`). This change the check such that one side must a
subset of the other, so if `x` is `FOO | BAR`, then `x == FOO` or `FOO == x` is
valid, but `x == BAZ` and `BAZ == x` are not.