Fix cast class error with join temporary
This fixes a bytecode generation bug causing invalid cast when a signature is
used on the right-hand side of a `LeftJoin`. This bug keeps popping up in this
code and it is difficult because the join has several similar but distinct
paths. In this case, I believe the issue is that the signature was already
being copied into the join temporary, but the caller was trying to recomputed
the signature on the join temporary (rather than the original value).