Show the base search on the actions page
Add a panel to display the base search (the one selected from the drop down)
filters. This was dropped for two reasons:

- in the reimplementation, it was more practical to have code to only draw the
  `BasicQuery` object rather than `ActionFilter` since that required reverse
  engineering the `BasicQuery` context from the `ActionFilter`, which was hard.
- I was the only one who ever looked at it

I want to see that information and it is simpler to have a separate rendering
for the base filters that is much simpler, since it's not editable.

In the refactor, the display logic for `BasicQuery` kept the old name of
`renderFilters`. This changes that to a more accurate `renderQuery` and adds a
new `renderFilters`.