Fix sequence generator to have no duplicates
The original design was to look back at the previously generated value to
ensure that when someone has to enter a sequence of random numbers, there are
no duplicates. However, the code was doing it to the input sequence (which was
initialised to all zeros) rather than the generated sequence.