Fix bug preventing importing `.niassawf` files in simulation
The fake action importer was incorrectly recognizing `.niassawf` files as its
own and importing them in a way that produced broken behaviour. This changes
the import process to check for `"kind": "action"`, which was added as part of
the new definitions page and then it validates all of the parameters as it
reads them. This also adds this property to exported actions so they round-trip
correctly.