Fix `Dump` in join operations
The dumpers are scoped to the olive, so when a `Dump` is included indirectly
throught a `Join` operation using a `Define`, the output from the dump
operation goes nowhere useful. This triggers dumper generation at the time of
olive creation so that the dumpers from `Define` will be attached to the
calling olive.