Allow dumpers to have column names
Dumpers require the original source to have meaningful column names, which is
frustrating during debugging and export. This changes the dumpers to have named
columns (similar to refillers). Olive writers can either choose a column name
or use an automatically inferred one.

This also fixes a bug in the simulation UI where the incorrect name was used
for the file name when downloading dumper output.