Attempt to fix LIMS key locking
There is an unhandled error condition where if the data being downloaded
produces an exception while processing, this will trigger the LIMS key locks to
be released even though the cache is still stale as it was never updated due to
the exception. This takes a more secure approach of downloading all the data
and then only releasing LIMS key locks for LIMS keys it sees in the data. The
streaming approach was meant to limit memory usage and while this approach will
use more memory, it should only duplicate the cost of the container (_i.e._,
the array list) rather than a duplicate cost of the objects themselves, so it
is still considerably less memory-intensive than the previous non-streaming
approach.