Deal with Niassa's IUS attributes being a lie
While the type for Niassa's IUS attribute looks like it combines the attributes
of any IUSes associated with a file, in fact, it does not. This is really bad
when multiple LIMS keys are associated with a single file. This change tries to
forcibly disable signatures on these runs and lets the actions fall into HALP
state.