Allow simulating existing olives
It is currently possible to simulate an existing olive but this relies on the
simulation dashboard being able to be populated with a particular file name
from a query parameter. This creates a new simulation endpoint that takes a
path to an existing script instead of a script itself.

This is for future feature support.