Add an ECMAScript 6 compilation target to olive expressions
This is part of a larger project to create guided meditations. They will run
partially on the front end and partially on the backend. To make them easier to
write, the goal is to have a single compiler that will generate front-end and
backend code. Not all of the existing olive language will be available on the
front end; only expressions. This change adds ES6 code generation to
`ExpressionNode` and all classes required by it and creates a new `runtime.ts`
which is analogous to `RuntimeSupport.java`