Allow saving action IDs from the UI
Add a new endpoint to get action IDs for a particular search. Since this
endpoint does not convert the actions to JSON, it is much cheaper than using
the normal action query end-point and extracting the action IDs. The goal of
this is to create a snapshot of actions for putting in tickets rather than a
query that can hide new actions.