Add new Cerberus plugin
This provides a new source of the `cerberus_fp` data format using Vidarr and
Pinery via the new Cerberus library. It adds a new data format,
`cerberus_error` for workflows that are were unable to be joined. This makes a
breaking change to Pinery configuration by changing the version from a string
to an integer. For compatibility with Vidarr, both the `pinery_ius` and
`cerberus_fp` are extended to include a new variable, `external_key` that
provide LIMS information in the format expected by Vidarr actions.