Fix import rules for constants and signatures
This fixes a bug where imports do not work for constants and signatures. This
means that `now` would fail to resolve to `std::date::now` even if `Import
std::date::*;` was present.

This occured because the actions, refillers, callable olives, and functions are
taken as a request from the point of use, so the rewrite rules can operate by
prepending the prefix to the name. Conversely, olives and signatures are fed
into the olive. The rewriting was applied, but lead to non-sense names like
`std::date::std::date::now`. This expands `ImportRewriter` to have a `strip`
method to rewrite `std::date::now` to `now`. However, when generating bytecode
`now` is no longer know, so `Target.unalisedName()` was added so
`std::date::now` will always be the name used during bytecode generation even
if the user specified `now` via an import.