Fix incorrect JavaScript generation in `Define`
A guidied meditation passes around all the variables in a JavaScript object and
each step takes this object as a single parameter. When a `GoTo` is invoked, it
expands the name parameters in the object to positional parameters. `Define`
needs to repack the positional parameters into an object. This adds the code to
generate the repacked object.