Fix stream capture in `Default`
For the code `x Default y`, Shesmu generates a call to `Optional.orElseGet`,
which takes a `Supplier` lambda. There is a "stream" value which is the current
row being iterated and variables that access the row read this variable. Some
lambdas take the steam as a capture (_i.e._, the stream is whatever the stream
outside) and others take the steam as a parameter (_i.e._, the stream is being
changed, such as in a grouping operation). The `Supplier` falls into the former
category but was generating a lambda of the latter type, resulting in incorrect
bytecode trying to access an argument that does not exist.