Allow setting metadata parameters globally for Vidarr
Vidarr workflows need to know how to associate the output files from the
workflow run to the LIMS metadata. Niassa had different handlers for different
cases; these were set in the `.niassawf` files as `"type"`, usually `FILES`.
Vidarr workflows require setting each output individually. For a typical
`FILES`-equivalent workflow, that might look:

```
metadata = {
  foo = ALL {output_prefix = "/srv/vidarr/storage/"},
  bar = ALL {output_prefix = "/srv/vidarr/storage/"},
  quux = ALL {output_prefix = "/srv/vidarr/storage/"}
}
```

This is tedious at best. So, the idea is to allow setting all of these values
simultaneously if they can be set that way. This changes the above to:

```
metadata = INDIVIDUAL {
  foo = ALL {output_prefix = "/srv/vidarr/storage/"},
  bar = ALL {output_prefix = "/srv/vidarr/storage/"},
  quux = ALL {output_prefix = "/srv/vidarr/storage/"}
}
```

or it can be simplified to:

```
metadata = GLOBAL { ALL {output_prefix = "/srv/vidarr/storage/"} }
```