Prevent plugin exceptions from breaking the _Actions_ page
The plugins can return a list of search to include on the _Actions_ page. If
one of them throws, the dashboard will fail to load. This logs the exceptions
but allows the dashboard to be built. This was in response to a JIRA bug where
the plugin would throw an exception if the login failed.