Change guided meditation implementation and fix bugs
The guided meditation `Flow By Match` had a bug:

```
Flow By Match
  When FOO {x} Then x # x was not available here
```

There were other inconsistent problems where variables that should be in scope
were not.

The ECMAScript code generator was originally a copy of the JVM implementation.
This lead to unnecessary complexity because the JVM implementation needs to
capture variables in lambdas while the ES6 does the capturing automatically.
The variables defined by the meditation were also placed in a state object that
the meditation console would pass between steps.

This change simplifies this by using the native capture system of ECMAScript
and removes the state object by moving that data into local variables. The
`Fetch`, `Form`, and `Fork` operations still need to be able to inject data
from the console, so they have a different signature where the console can pass
an object with named parameters that they can extract as necessary.

This removes generated ECMAScript lambdas from most `WizardNode` classes and
removes lambdas for loading variable names in much of the compiler code.

This fixes the variable capture in `Flow By Match` and implements it in
`Match`, where that was incomplete.

Also fixes a bug in `If` where the false branch was not rendered correctly.