Fix two compiler bugs in `For` expression
The first bug is that `Dict` collector is using the wrong type. The dictionary
generated is a `TreeMap` using a custom comparator. The comparator should match
the type of the key, but the collector was using the type of the items being
iterated over.

The second is a bug selecting free variables. In an expression like this:

     For x In xs: Let y = x + z List y

Consider `x + z`. `x` is coming from the `For` condition and `z` is coming from
outside this expression (_i.e._, it is free). Any list transformation (`Let`,
`Where`, `Flatten`) must do this name manipulation. Because some
transformations change the name (_e.g_, `Let`), each list transformation is
given a name and must produce a name for the downstream elements. Most
operations (_e.g._, `Where`) pass this name through unchanged.

The bug occurs because the names removed were *output* names rather than
*input* names. This would trigger a bug when a variable that should be free was
also an output name. The transformation would incorrectly identify it as bound
and not request it from the parent context. When trying to generate code, it
would then fail to find this variable.