Allow using custom lookup environment for input formats
Shesmu uses reflection and annotation to allow the compiler to bind to custom
data formats. When in a module, either the data format needs to be public or
the data format needs to provide a method handle lookup context with private
visibility. This allows the data format author to decide and moves some of the
internal data formats to private access.