Make `First` and `Reduce` order-sensitive and optionals ordered
This fixes a bug where `Reduce` and `First` can receive values in arbitrary
order. This forces values to be sorted. It may break existing code, but that
code is already broken.

Since optional values, when used as a source of iteration, contain only one
item, they should be considered ordered. This required changing the ordering
checks to occur after the type checks, since the type is required to know if
its an optional or list iteration.