Allow `Match` to operate on optional types
This does a subtle rewrite of an optional type `` ` `` _x_ `` ` `` to an
algebraic type of `NONE | SOME {` _x_ `}`.

This functionality is already covered using the `?` and `Default` operations,
but this syntax may be easier to read. The following two examples should be
interchangeable.

```
  Match x
    When NONE Then g()
    When SOME {y} Then f(y)
```

and

```
   `f(y?)` Default g()
```

The preference for which depends on the complexity of the intermediate
expressions.