Include detailed counts when running commands
When a command is executed, the server counts the number of actions that
responded to the command. This can be misleading since actions can
ignore the command. Additionally, some commands can lead to the action
purging, so the count of actions after refreshing will not make much
sense.

This changes the interface to report the number of actions matching the
search that did and did not respond to the command and the subset of
matching actions that want to be purged.