Prepare compiler for modularization
The compiler currently emits classes that are incompatible with the Java 9
Jigsaw module system. This change refactors the compiler so that it generates
module-compliant code in preparation for modularization. It does this by:

- making sure that all generated classes are in a single package
- adding a new method to `ActionGenerator` to provide a `MethodHandles.Lookup` instance

Under modularization, classes cannot reflect into classes outside of their own
module visibility. For generated code, this makes it impossible for the
compiler to rummage around in the class it just generated. To circumvent this,
the generated class must create a `Lookup` instance with the generated class's
visibility rules and then share this with the compiler.