Fix an aliasing bug when using `Prefix` and `Call` in a join
This corrects a bug where the inner key function of a `Call` operation needs to
use the original name for a stream variable even though it has been renamed
using `Prefix`. Meanwhile, the collectors need to use the aliased name since
they are interacting with a join temporary and not the original value. This was
not a problem for prefixed input formats since input formats provide their own
loaders.